requirements.txt
kornia==0.3.1
numpy==1.18.2
opencv-python-headless==4.2.0.34
Pillow==7.0.0
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.16.2
scipy==1.4.1
tensorboard-pytorch==0.7.1
torch==1.5.0
torchvision==0.6.0
tqdm==4.44.1
wandb==0.9.5
ipython==7.16.1
tensorboard==2.2.2
statsmodels==0.11.1
pytorch-fid==0.1.1
imgaug==0.4.0
